requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6

SQLAlchemy>=2.0.35
psycopg2-binary>=2.9.9
alembic>=1.14.0

pydantic>=2.9.2
pydantic-settings>=2.6.1
python-dotenv>=1.0.1

vosk>=0.3.45
websockets>=12.0